# specs/005-mcp-agent-tools/contracts/openapi.yml
openapi: 3.0.0
info:
  title: MCP Task Tools API
  description: API endpoints that represent the tools available to the AI agent for task management.
  version: 1.0.0

paths:
  /tasks:
    get:
      summary: List tasks (list_tasks tool)
      operationId: list_tasks
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, completed]
            default: open
          description: The status of tasks to retrieve.
      responses:
        '200':
          description: A list of tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a task (add_task tool)
      operationId: add_task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: The created task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

  /tasks/{task_id}:
    patch:
      summary: Update a task (update_task tool)
      operationId: update_task
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: The updated task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      summary: Delete a task (delete_task tool)
      operationId: delete_task
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted successfully.

  /tasks/{task_id}/complete:
    post:
      summary: Complete a task (complete_task tool)
      operationId: complete_task
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The completed task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, completed]

    TaskCreate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string